Fixing Edit Book
So that heading shows up as Edit Book with book title when on Edit page rather than new book etc.  Also got rid of list in alerts, cus it looked weird. Gotta find a way of making them green when they're good and red when they're bad/ tick or x.